+++ lib/dasboard_screens/pc_builder/models/details.dart
@@ -1,6 +1,6 @@
 import 'package:carousel_slider/carousel_slider.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:fblogin/dasboard_screens/pc_builder/collection_page.dart';
+import 'package:fblogin/dasboard_screens/pc_builder/pages/collection_page.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
